pyproject.toml
[tool.poetry]
name = "from-csv-to-db"
version = "0.1.0"
description = ""
authors = ["maciej88"]
readme = "README.md"
packages = [{include = "from_csv_to_db"}]

[tool.poetry.dependencies]
python = "^3.11"
alembic = "^1.11.1"
python-dotenv = "^1.0.0"
loguru = "^0.7.0"
psycopg2-binary = "^2.9.6"
pandas = "^2.0.3"
asyncpg = "^0.28.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
